TS: extracting List from pages/Set
